refactor(dice): merge dice-slice1.test.ts into DamageRoll.test.ts
The slice1 file was a slice-1 artifact: the original DamageRoll.test.ts
had no evaluation coverage, so slice 1 shipped new evaluation tests in
a separate file to avoid touching the existing suite. That reason is
gone and the "slice1" name was development-process taxonomy leaking
into the codebase.

Move all 19 tests into DamageRoll.test.ts under two new describe blocks:
- 'DamageRoll evaluation — primary die outcome' (regression + bug5)
- 'DamageRoll integration — ItemActivationManager AoE and proficiency'
  (bug6, bug7, bug8b)

Port the evaluation helpers (stubBaseRollEvaluate, stagePrimaryDieResults,
ensureGameUserTargets, makeMockDamageRollClass, makeAoEItem, DieResult).
DamageRoll.test.ts had no prior evaluation setup, so no collisions.

Test names preserved. No assertion changes. 731/731 green.
